+++ comp2013/src/main/java/comp2013/SnakeImageUtil.java
@@ -32,6 +32,12 @@ public class SnakeImageUtil
 		m_SnakeImages.put("cloud-background", new Image(SnakeImageUtil.class.getResource
 				("/images/cloud-background.png").toExternalForm()));
 
+		m_SnakeImages.put("grass-background", new Image(SnakeImageUtil.class.getResource
+				("/images/grass-background.jpg").toExternalForm()));
+
+		m_SnakeImages.put("ocean-background", new Image(SnakeImageUtil.class.getResource
+				("/images/ocean-background.jpg").toExternalForm()));
+
 		m_SnakeImages.put("jungle-background", new Image(SnakeImageUtil.class.getResource
 				("/images/jungle-background.jpg").toExternalForm()));